pyproject.toml
[tool.poetry]
name = "python-drf"
version = "0.1.0"
description = ""
authors = ["VKnyshov <valeriy.knyshov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.6"
django-dotenv = "^1.4.2"
python-decouple = "^3.8"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
